converted the rest of the fun commands
